Implement App Scaffold
Utilized Logo, Timer, and GithubLink

These three components are rendered in scaffold as required to fit the
zeplin mockup.

Currently only accepts children prop but in the future should
accept props that are used in Logo and Timer as well.

In home.scss I added a background color to the whole body

In reset.scss I added 100% height to root, html, and body so that the page
renders the scaffold at 100%

Currently index.js is rendering App which only returns the scaffold.

May want to split App into a separate file, but this works for now.