Add Timer functionality
Change App to return TestPage component

Pass into scaffold minutes and seconds props for the Timer

Set a time state with 1 minutes and 0 seconds initially, set an isActive
state to check if timer is active, and have a timerId state for the
setInterval.

UseEffect checks if timer is active and if it is setIntervals a tick
function. As a cleanup we clearInterval if the timer is
not active but there is more seconds left and as a return.

Tick checks what time.s is. On the initial tick, we set time.s to 59
otherwise we time.s - 1. We also clearInterval and setIsActive to false
when the time runs out.

Restart simply sets the time to 1 minute and 0 seconds and sets the
timer to inactive, as well as clears the interval.

Toggle is how we start the timer as well as whether the buttons are
disabled.