Create Button (#41)
* Create Button component

Created a button that accepts onClick event toggle, as expressed in
TestPage.jsx.

The button is disabled at only one point, when the game has not started
and the buttons name is pause. I have a useEffect checking if the game
isActive and if the button is named pause, this controls the buttons
disabledness. I clean up the useEffect with a return callback.

To check if there is an icon, i check if the iconName being passed in
is default. Default is used for both start test and play again (which
are expressed differently in text).

The icon is in a span with a preset margin to push the text over. This
is to account for the default, where we want the text to be centered.

In Icon.jsx I added to prop-types name default, as this can now be
passed to help account for start test and play again.

The storybook should account for all cases, so play around with that and
let me know if anything comes up.

I also changed in typography the font weight for button to be medium as
outlined in the design.

* change to 0

* better understanding of disabled

* change around classname

* aria label make more sense?