Make initial input design
Implemented a simple input designed on initial design.

At the moment it is limited to 4 characters as this is how many
characters fit into the input/ how many we can expect.

It accepts on function onKeyDown which will be defined in TestPage,
where the input will be used.

There is one problem with our storybook accessibility which will need to
be fixed later.

Also created new mixins for the input text as inputs do not inherently
inherent