Implement buttons on testpage (#42)
* Implement buttons on testpage

Made a temporary input with a function startTimer(). This is only
temporary and will be cleaned up later.

Currently the button functions the way it does through a new state
isPauseButtonShowing. This state dictates whether to show the pause
button or to show the start/restart buttons. It does so in the toggle
function. It just sets pauseButtonShowing to the opposite of isActive.

* implement ternary to clean up code