Implement input block
Created a story to host the input alongside completed words and words
todo.

The general idea is that we have words lined up in a div to the right
of the input, and onsubmit of the input the word is moved
to a div on the left of the input called completed. Other fancy things,
happen such as color and strikethroughs. See sass for this.

WordsTodo state will be passed in later from a different file, but
at the moment are just a test with dog.