Implement InputBlock (#45)
* Implement InputBlock

After realizing that we needed to reset the words, I decided to clean
up the code into a useReducer function. This includes and intial state,
and a switch that changes the "state" of anything in the intial state.
I dispatch commands/actions to this reducer and update the state as needed.

For example, on the case of pause I update the state of isActive, inputDisabled,
and isPauseShowing. These dispatches should be the only heavy logic, the
other parts include control flow as to what to dispatch and inputRef
logic/Clear interval logic.

I cannot at the moment figure out to return focus to the input on restart/
start button.

I also made it so the words to submitted automatically if the current value
of the input is the same as the answer. I do repeat myself here (in the
startTimer() line 153-164). I had to change several files to include a
onKeyUp. This is because we had it as onKeyDown and the current value
was not updating quickly on the same level of where the user is when typing
the word. In other words if a user types 'dog' they will type 'd' but
the current value is '' onKeyDown. OnKeyUp it would be registered on the
key up and be 'd'. Useful.

* Reduce booleans and Create a appState instead

Created a new object with 4 app states. Utilized these appstates
to change what to render instead of using previous 4 booleans.

In my dispatches I update the appstate key from my state to match what
the app should be doing, ie play, pause, initialize, or be finished. I
also created 4 variables to reduce the clutter of comparisons in my
return jsx.

* Clean up submit function

Made the function only dispatch one type with an action to decide
what isCorrect should be

* make submitWord function to DRY a bit

* add changes also added .storybook directory to gitignore