Implement basic accuracy component
Utilizing a dashoffset and a math equation (I don't know what it is
doing exactly) I calculate how far the stroke in an svg would need to
be to be a certain percentage. I have two other paths
1. Is the background half circle
2. Is the white outline
3. Is the middle gradient

The gradient is a rainbow gradient which is certain colors bleeding
into other colors at certain percentages.