Implement Kana component and design Input (#50)
Input component needed to have a div wrap both the current word and
the input. They are displayed next to each other with display flex and
this wrapper has a position of relative with a width pre-made. The Input
class stays the same however the current word is given a position of
absolute and set right 0 to have it fully right.

InputBlock needed the Kana component in the maps, which required some
logic. They are also given a margin to keep some distance between them.
The return has some logic to know whether to show test paused or the
input block. This required a fragment which our linter makes into <></>

Our Kana component needed a margin so I added it to the sass file and
the classnames.

Our scaffold needed children updated to node

In Testpage I pass the pauseState to the inputblock to know whether
to show testpaused. I also added to the useeffect
inputRef.current.focus(); so on the push of play our input is
focused. I also fixed a bug, when a user pushed tab they started the
timer because of the onKeyPress. I just added some logic to solve this.
The basic premise being if the keypressed is not a letter or enter
dont go forward with the timer. I also added the instructions.