Fix Bugs
Two/three bugs where happening.

First two where more or less the same bug. I had inputFocus in the useEffect. This was making it impossible to tab onto the pause and if you manage to pause it, when you played again the timer was messed up. I fixed this by adding a new state. In the useeffect I check for this state and focus on the input only this one time as I clean it up and set the state to false.

The third/second major bug was when a user pressed enter to submit a word the timer would reset. I added another bit to the control flow to check if the use pressed enter and the time was 1 only to set the timer to 59 seconds left.