Homepage (#58)
* Add variant prop to scaffold

Homepage will use scaffold but does not need the timer. Added variant to make sure timer does not show up on homepage.

* Make home page component

Added a few props to change colors in of components that are different from testpage. Also added a tempory div around the scaffold for testpage. Added some margins, need to work on this part.

* Finalize Homepage

Added routing to webpack

Edit around styles

Co-authored-by: Joshuah Martin <joshuahmartin@Joshuahs-MacBook-Pro.local>