J main (#67)
* Homepage (#58)

* Add variant prop to scaffold

Homepage will use scaffold but does not need the timer. Added variant to make sure timer does not show up on homepage.

* Make home page component

Added a few props to change colors in of components that are different from testpage. Also added a tempory div around the scaffold for testpage. Added some margins, need to work on this part.

* Finalize Homepage

Added routing to webpack

Edit around styles

Co-authored-by: Joshuah Martin <joshuahmartin@Joshuahs-MacBook-Pro.local>

* Accuracy (#59)

* Implement basic accuracy component

Utilizing a dashoffset and a math equation (I don't know what it is
doing exactly) I calculate how far the stroke in an svg would need to
be to be a certain percentage. I have two other paths
1. Is the background half circle
2. Is the white outline
3. Is the middle gradient

The gradient is a rainbow gradient which is certain colors bleeding
into other colors at certain percentages.

* add score mixin for the accuracy meter.

Co-authored-by: Joshuah Martin <joshuahmartin@Joshuahs-MacBook-Pro.local>